pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asbg"
version = "0.0.2"
authors = [
  { name="Thomas Erard", email="erard.thomas@gmail.com" },
]
description = "Display the results of the ASBG players in competition"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "beautifulsoup4",
  "click",
  "dash-bootstrap-components",
  "dash",
  "google-api-python-client",
  "google-auth-oauthlib",
  "pandas",
  "polars",
  "PyYAML",
  "requests",
  "streamlit",
  "tabulate",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest",
]

[project.scripts]
# Defines one entrypoint: a click command called "cli" in the module asbg/cli.py
asbg = "asbg.cli:cli"

[tool.pytest]
cache_dir  = ".cache"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]

[tool.ruff]
cache-dir = ".cache"
line-length = 100
lint.extend-select = ["I"]

[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["asbg"]
